examples/post_training_quantization/torch/save_load_mobilenet_v2/requirements.txt
fastdownload==0.0.7
openvino==2024.1.0
scikit-learn==1.2.2
torch==2.2.1
torchvision==0.17.1
onnx==1.16.0
onnxruntime-openvino==1.16.0
